refactor-#156: move customer mutation to mutations
