+++ src/microservices/customer/graphql/mutations/index.ts
@@ -1,0 +1,1 @@
+export * from './customer-mutation.resolver';
